pyproject.toml

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "agentic-fabriq-sdk"
version = "0.1.25"
description = "Agentic Fabriq SDK: high-level client, CLI tool, DX helpers, and auth for AI agents"
readme = "README.md"
license = "Apache-2.0"
authors = ["Agentic Fabriq Contributors <contributors@agentic-fabriq.org>"]
packages = [
    { include = "af_sdk", from = "." },
    { include = "af_cli", from = "." }
]
keywords = ["fabriq", "agentic-fabriq", "sdk", "ai", "agents", "agentic", "fabric", "cli"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
include = ["af_sdk/py.typed"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
# Core SDK dependencies
httpx = ">=0.25.0"
pydantic = ">=2.4.0"
typing-extensions = ">=4.0.0"
PyJWT = ">=2.8.0"
# OpenTelemetry for tracing
opentelemetry-api = ">=1.20.0"
opentelemetry-instrumentation-httpx = ">=0.41b0"
# NATS for event streaming
nats-py = ">=2.4.0"
# Additional HTTP client support
aiohttp = ">=3.9.0"
# Plugin/extension system
stevedore = ">=5.1.0"
# CLI dependencies
typer = {extras = ["all"], version = ">=0.20.0"}
rich = ">=13.7.0"
keyring = ">=25.0.0"
cryptography = ">=41.0.0"
click = ">=8.1.0"
PyYAML = ">=6.0.0"

[tool.poetry.scripts]
afctl = "af_cli.main:app"

[tool.poetry.urls]
Homepage = "https://github.com/agentic-fabric/agentic-fabric"
Repository = "https://github.com/agentic-fabric/agentic-fabric"
Documentation = "https://docs.agentic-fabric.org"
